Added validation to make sure user is returning bike to valid station. If user returns bike after rental end_time, set over_time = true.
Not sure if this is the best location for over_time = true? Basically just when the bike is returned I say is it past your rental end_time? and if it is I update the bool to be true. Totally happy to chat about this and move it in the future though.